+++ helm-projectile.el
@@ -40,10 +40,12 @@
 ;;; Code:
 
 (require 'projectile)
+(require 'cl-lib)
+(require 'helm)
+(require 'helm-types)
 (require 'helm-locate)
 (require 'helm-buffers)
 (require 'helm-files)
-(require 'cl-lib)
 
 (declare-function eshell "eshell")
 (declare-function helm-do-ag "helm-ag")
@@ -145,6 +147,20 @@ DIR is the project root."
         (default-directory dir))
     (projectile-compile-project helm-current-prefix-arg dir)))
 
+(defun helm-projectile-test-project (dir)
+  "A Helm action for test a project.
+DIR is the project root."
+  (let ((helm--reading-passwd-or-string t)
+        (default-directory dir))
+    (projectile-test-project helm-current-prefix-arg dir)))
+
+(defun helm-projectile-run-project (dir)
+  "A Helm action for run a project.
+DIR is the project root."
+  (let ((helm--reading-passwd-or-string t)
+        (default-directory dir))
+    (projectile-run-project helm-current-prefix-arg dir)))
+
 (defun helm-projectile-remove-known-project (_ignore)
   "Delete selected projects.
 _IGNORE means the argument does not matter.
@@ -154,7 +170,7 @@ It is there because Helm requires it."
     (with-helm-display-marked-candidates
       helm-marked-buffer-name
       projects
-      (if (not (y-or-n-p (format "Delete *%s projects(s)" len)))
+      (if (not (y-or-n-p (format "Delete *%s projects(s)? " len)))
           (message "(No deletion performed)")
         (progn
           (mapc (lambda (p)
@@ -162,6 +178,21 @@ It is there because Helm requires it."
                 projects)
           (projectile-save-known-projects))
         (message "%s projects(s) deleted" len)))))
+
+(defvar helm-projectile-projects-map
+  (let ((map (make-sparse-keymap)))
+    (set-keymap-parent map helm-map)
+    (helm-projectile-define-key map
+      (kbd "C-d") #'dired
+      (kbd "M-g") #'helm-projectile-vc
+      (kbd "M-e") #'helm-projectile-switch-to-eshell
+      (kbd "C-s") #'helm-find-files-grep
+      (kbd "M-c") #'helm-projectile-compile-project
+      (kbd "M-t") #'helm-projectile-test-project
+      (kbd "M-r") #'helm-projectile-run-project
+      (kbd "M-D") #'helm-projectile-remove-known-project)
+    map)
+  "Mapping for known projectile projects.")
 
 (defvar helm-source-projectile-projects
   (helm-build-in-buffer-source "Projectile projects"
@@ -171,17 +202,8 @@ It is there because Helm requires it."
                       (projectile-relevant-known-projects))
               projectile-known-projects))
     :fuzzy-match helm-projectile-fuzzy-match
-    :keymap (let ((map (make-sparse-keymap)))
-              (set-keymap-parent map helm-map)
-              (helm-projectile-define-key map
-                (kbd "C-d") 'dired
-                (kbd "M-g") 'helm-projectile-vc
-                (kbd "M-e") 'helm-projectile-switch-to-eshell
-                (kbd "C-s") 'helm-find-files-grep
-                (kbd "M-c") 'helm-projectile-compile-project
-                (kbd "M-D") 'helm-projectile-remove-known-project)
-              map)
-    :mode-line helm-ff-mode-line-string
+    :keymap helm-projectile-projects-map
+    :mode-line helm-read-file-name-mode-line-string
     :action '(("Switch to project" .
                (lambda (project)
                  (let ((projectile-completion-system 'helm))
@@ -242,7 +264,8 @@ It is there because Helm requires it."
                    (buffer-list))))
 
 (defvar helm-projectile-virtual-dired-remote-enable nil
-  "Enable virtual Dired manager in remote host. Disabled by default.")
+  "Enable virtual Dired manager on remote host.
+Disabled by default.")
 
 (defun helm-projectile-dired-files-new-action (candidate)
   "Create a Dired buffer from chosen files.
@@ -312,7 +335,7 @@ CANDIDATE is the selected file.  Used when no file is explicitly marked."
                                                           files)
                                                 (list candidate))))
               (rename-buffer dired-buffer-name))))
-      (error "You're not in a Dired buffer to add."))))
+      (error "You're not in a Dired buffer to add"))))
 
 (defun helm-projectile-dired-files-delete-action (candidate)
   "Delete selected entries from a Dired buffer.
@@ -343,20 +366,19 @@ CANDIDATE is the selected file.  Used when no file is explicitly marked."
 (defvar helm-projectile-find-file-map
   (let ((map (copy-keymap helm-find-files-map)))
     (helm-projectile-define-key map
-      (kbd "C-c f") 'helm-projectile-dired-files-new-action
-      (kbd "C-c a") 'helm-projectile-dired-files-add-action
-      (kbd "M-e") 'helm-projectile-switch-to-eshell
-      (kbd "M-.") 'helm-projectile-ff-etags-select-action
-      (kbd "M-!") 'helm-projectile-find-files-eshell-command-on-file-action)
-    (define-key map (kbd "<left>") 'helm-previous-source)
-    (define-key map (kbd "<right>") 'helm-next-source)
+      (kbd "C-c f") #'helm-projectile-dired-files-new-action
+      (kbd "C-c a") #'helm-projectile-dired-files-add-action
+      (kbd "M-e") #'helm-projectile-switch-to-eshell
+      (kbd "M-.") #'helm-projectile-ff-etags-select-action
+      (kbd "M-!") #'helm-projectile-find-files-eshell-command-on-file-action)
+    (define-key map (kbd "<left>") #'helm-previous-source)
+    (define-key map (kbd "<right>") #'helm-next-source)
     map)
   "Mapping for file commands in Helm Projectile.")
 
 (defvar helm-projectile-file-actions
   (helm-projectile-hack-actions
-   (cdr (assq 'action (or helm-source-find-files
-                          (helm-make-source "Find files" 'helm-source-ffiles))))
+   helm-find-files-actions
    ;; Delete these actions
    'helm-ff-browse-project
    'helm-insert-file-name-completion-at-point
@@ -386,7 +408,7 @@ CANDIDATE is the selected file.  Used when no file is explicitly marked."
     :action-transformer 'helm-find-files-action-transformer
     :keymap helm-projectile-find-file-map
     :help-message helm-ff-help-message
-    :mode-line helm-ff-mode-line-string
+    :mode-line helm-read-file-name-mode-line-string
     :action helm-projectile-file-actions))
 
 (defvar helm-source-projectile-files-list
@@ -399,7 +421,7 @@ CANDIDATE is the selected file.  Used when no file is explicitly marked."
     :coerce 'helm-projectile-coerce-file
     :keymap helm-projectile-find-file-map
     :help-message 'helm-ff-help-message
-    :mode-line helm-ff-mode-line-string
+    :mode-line helm-read-file-name-mode-line-string
     :action helm-projectile-file-actions
     )
   "Helm source definition for Projectile files.")
@@ -414,7 +436,7 @@ CANDIDATE is the selected file.  Used when no file is explicitly marked."
     :coerce 'helm-projectile-coerce-file
     :keymap helm-find-files-map
     :help-message 'helm-ff-help-message
-    :mode-line helm-ff-mode-line-string
+    :mode-line helm-read-file-name-mode-line-string
     :action helm-projectile-file-actions
     )
   "Helm source definition for all Projectile files in all projects.")
@@ -448,7 +470,7 @@ CANDIDATE is the selected file.  Used when no file is explicitly marked."
                 (kbd "C-c d") 'helm-projectile-dired-files-delete-action)
               map)
     :help-message 'helm-ff-help-message
-    :mode-line helm-ff-mode-line-string
+    :mode-line helm-read-file-name-mode-line-string
     :action helm-projectile-dired-file-actions)
   "Helm source definition for Projectile delete files.")
 
@@ -476,16 +498,16 @@ CANDIDATE is the selected file.  Used when no file is explicitly marked."
     :keymap (let ((map (make-sparse-keymap)))
               (set-keymap-parent map helm-map)
               (helm-projectile-define-key map
-                (kbd "<left>") 'helm-previous-source
-                (kbd "<right>") 'helm-next-source
-                (kbd "C-c o") 'helm-projectile-dired-find-dir-other-window
-                (kbd "M-e")   'helm-projectile-switch-to-eshell
-                (kbd "C-c f") 'helm-projectile-dired-files-new-action
-                (kbd "C-c a") 'helm-projectile-dired-files-add-action
-                (kbd "C-s")   'helm-find-files-grep)
+                (kbd "<left>") #'helm-previous-source
+                (kbd "<right>") #'helm-next-source
+                (kbd "C-c o") #'helm-projectile-dired-find-dir-other-window
+                (kbd "M-e")   #'helm-projectile-switch-to-eshell
+                (kbd "C-c f") #'helm-projectile-dired-files-new-action
+                (kbd "C-c a") #'helm-projectile-dired-files-add-action
+                (kbd "C-s")   #'helm-find-files-grep)
               map)
     :help-message 'helm-ff-help-message
-    :mode-line helm-ff-mode-line-string
+    :mode-line helm-read-file-name-mode-line-string
     :action '(("Open Dired" . helm-projectile-dired-find-dir)
               ("Open Dired in other window `C-c o'" . helm-projectile-dired-find-dir)
               ("Switch to Eshell `M-e'" . helm-projectile-switch-to-eshell)
@@ -520,7 +542,6 @@ CANDIDATE is the selected file.  Used when no file is explicitly marked."
    (persistent-action :initform 'helm-buffers-list-persistent-action)
    (keymap :initform helm-buffer-map)
    (volatile :initform t)
-   (mode-line :initform helm-buffer-mode-line-string)
    (persistent-help
     :initform
     "Show this buffer / C-u \\[helm-execute-persistent-action]: Kill this buffer")))
@@ -537,7 +558,7 @@ CANDIDATE is the selected file.  Used when no file is explicitly marked."
     :coerce 'helm-projectile-coerce-file
     :keymap helm-projectile-find-file-map
     :help-message 'helm-ff-help-message
-    :mode-line helm-ff-mode-line-string
+    :mode-line helm-read-file-name-mode-line-string
     :action helm-projectile-file-actions
     )
   "Helm source definition for recent files in current project.")
@@ -623,7 +644,7 @@ Other file extensions can be customized with the variable `projectile-other-file
                                        (define-key map (kbd "<right>") 'helm-next-source)
                                        map)
                              :help-message helm-ff-help-message
-                             :mode-line helm-ff-mode-line-string
+                             :mode-line helm-read-file-name-mode-line-string
                              :action helm-projectile-file-actions)
                   :buffer "*helm projectile*"
                   :prompt (projectile-prepend-project-name "Find other file: ")))))
@@ -636,13 +657,14 @@ ACK-IGNORED-PATTERN is a file regex to exclude from searching.
 ACK-EXECUTABLE is the actual ack binary name.
 It is usually \"ack\" or \"ack-grep\".
 If it is nil, or ack/ack-grep not found then use default grep command."
+  (require 'grep)
   (let* ((default-directory (projectile-project-root))
          (helm-ff-default-directory (projectile-project-root))
          (follow (and helm-follow-mode-persistent
                       (assoc-default 'follow helm-source-grep)))
          (helm-grep-in-recurse t)
-         (grep-find-ignored-files (-union (projectile-ignored-files-rel)  grep-find-ignored-files))
-         (grep-find-ignored-directories (-union (projectile-ignored-directories-rel) grep-find-ignored-directories))
+         (helm-grep-ignored-files (-union (projectile-ignored-files-rel)  grep-find-ignored-files))
+         (helm-grep-ignored-directories (-union (projectile-ignored-directories-rel) grep-find-ignored-directories))
          (helm-grep-default-command (if use-ack-p
                                         (concat ack-executable " -H --no-group --no-color " ack-ignored-pattern " %p %f")
                                       "grep -a -r %e -n%cH -e %p %f ."))
@@ -659,7 +681,6 @@ If it is nil, or ack/ack-grep not found then use default grep command."
             :filter-one-by-one 'helm-grep-filter-one-by-one
             :candidate-number-limit 9999
             :nohighlight t
-            :mode-line helm-grep-mode-line-string
             ;; We need to specify keymap here and as :keymap arg [1]
             ;; to make it available in further resuming.
             :keymap helm-grep-map
@@ -738,14 +759,15 @@ If it is nil, or ack/ack-grep not found then use default grep command."
   (interactive (if current-prefix-arg (list (read-string "option: " "" 'helm-ag-command-history))))
   (if (require 'helm-ag nil  'noerror)
       (if (projectile-project-p)
-          (let* ((grep-find-ignored-files (-union (projectile-ignored-files-rel)  grep-find-ignored-files))
+          (let* ((grep-find-ignored-files (-union (projectile-ignored-files-rel) grep-find-ignored-files))
                  (grep-find-ignored-directories (-union (projectile-ignored-directories-rel) grep-find-ignored-directories))
                  (ignored (mapconcat (lambda (i)
                                        (concat "--ignore " i))
                                      (append grep-find-ignored-files grep-find-ignored-directories)
                                      " "))
                  (helm-ag-command-option options)
-                 (helm-ag-base-command (concat helm-ag-base-command " " ignored)))
+                 (helm-ag-base-command (concat helm-ag-base-command " " ignored))
+                 (current-prefix-arg nil))
             (helm-do-ag (projectile-project-root)))
         (error "You're not in a project"))
     (error "helm-ag not available")))
@@ -787,31 +809,31 @@ If it is nil, or ack/ack-grep not found then use default grep command."
   "Toggle Helm version of Projectile commands."
   (if (> toggle 0)
       (progn
-        (setq projectile-switch-project-action 'helm-projectile-find-file)
-        (define-key projectile-command-map (kbd "a") 'helm-projectile-find-other-file)
-        (define-key projectile-command-map (kbd "f") 'helm-projectile-find-file)
-        (define-key projectile-command-map (kbd "F") 'helm-projectile-find-file-in-known-projects)
-        (define-key projectile-command-map (kbd "g") 'helm-projectile-find-file-dwim)
-        (define-key projectile-command-map (kbd "d") 'helm-projectile-find-dir)
-        (define-key projectile-command-map (kbd "p") 'helm-projectile-switch-project)
-        (define-key projectile-command-map (kbd "e") 'helm-projectile-recentf)
-        (define-key projectile-command-map (kbd "b") 'helm-projectile-switch-to-buffer)
-        (define-key projectile-command-map (kbd "s g") 'helm-projectile-grep)
-        (define-key projectile-command-map (kbd "s a") 'helm-projectile-ack)
-        (define-key projectile-command-map (kbd "s s") 'helm-projectile-ag)
+        (setq projectile-switch-project-action #'helm-projectile-find-file)
+        (define-key projectile-command-map (kbd "a") #'helm-projectile-find-other-file)
+        (define-key projectile-command-map (kbd "f") #'helm-projectile-find-file)
+        (define-key projectile-command-map (kbd "F") #'helm-projectile-find-file-in-known-projects)
+        (define-key projectile-command-map (kbd "g") #'helm-projectile-find-file-dwim)
+        (define-key projectile-command-map (kbd "d") #'helm-projectile-find-dir)
+        (define-key projectile-command-map (kbd "p") #'helm-projectile-switch-project)
+        (define-key projectile-command-map (kbd "e") #'helm-projectile-recentf)
+        (define-key projectile-command-map (kbd "b") #'helm-projectile-switch-to-buffer)
+        (define-key projectile-command-map (kbd "s g") #'helm-projectile-grep)
+        (define-key projectile-command-map (kbd "s a") #'helm-projectile-ack)
+        (define-key projectile-command-map (kbd "s s") #'helm-projectile-ag)
         (helm-projectile-commander-bindings))
     (progn
-      (setq projectile-switch-project-action 'projectile-find-file)
-      (define-key projectile-command-map (kbd "a") 'projectile-find-other-file)
-      (define-key projectile-command-map (kbd "f") 'projectile-find-file)
-      (define-key projectile-command-map (kbd "F") 'projectile-find-file-in-known-projects)
-      (define-key projectile-command-map (kbd "g") 'helm-projectile-find-file-dwim)
-      (define-key projectile-command-map (kbd "d") 'projectile-find-dir)
-      (define-key projectile-command-map (kbd "p") 'projectile-switch-project)
-      (define-key projectile-command-map (kbd "e") 'projectile-recentf)
-      (define-key projectile-command-map (kbd "b") 'projectile-switch-to-buffer)
-      (define-key projectile-command-map (kbd "s g") 'projectile-grep)
-      (define-key projectile-command-map (kbd "s s") 'projectile-ag)
+      (setq projectile-switch-project-action #'projectile-find-file)
+      (define-key projectile-command-map (kbd "a") #'projectile-find-other-file)
+      (define-key projectile-command-map (kbd "f") #'projectile-find-file)
+      (define-key projectile-command-map (kbd "F") #'projectile-find-file-in-known-projects)
+      (define-key projectile-command-map (kbd "g") #'helm-projectile-find-file-dwim)
+      (define-key projectile-command-map (kbd "d") #'projectile-find-dir)
+      (define-key projectile-command-map (kbd "p") #'projectile-switch-project)
+      (define-key projectile-command-map (kbd "e") #'projectile-recentf)
+      (define-key projectile-command-map (kbd "b") #'projectile-switch-to-buffer)
+      (define-key projectile-command-map (kbd "s g") #'projectile-grep)
+      (define-key projectile-command-map (kbd "s s") #'projectile-ag)
       (projectile-commander-bindings))))
 
 ;;;###autoload
@@ -833,7 +855,7 @@ If invoked outside of a project, displays a list of known projects to jump."
 ;;;###autoload
 (eval-after-load 'projectile
   '(progn
-     (define-key projectile-command-map (kbd "h") 'helm-projectile)))
+     (define-key projectile-command-map (kbd "h") #'helm-projectile)))
 
 (provide 'helm-projectile)